Moving to an stable version of the Monitor API's
