+++ azurerm/resource_arm_metric_alertrule.go
@@ -3,10 +3,9 @@ package azurerm
 import (
 	"fmt"
 	"log"
-
 	"strings"
 
-	"github.com/Azure/azure-sdk-for-go/services/monitor/mgmt/2017-05-01-preview/insights"
+	"github.com/Azure/azure-sdk-for-go/services/monitor/mgmt/2018-03-01/insights"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
